+++ hamb/handlers/sample_test_failed_handler.py
@@ -1,5 +1,5 @@
 """
-this is a test handler for the unit test - test_ham_run
+this is a test handler for the unit test - test_ham_run.
 """
 
 from datacoco_core import Logger