+++ hamb/handlers/slack_handler.py
@@ -1,11 +1,15 @@
-"""
-here is where the slack handler will go
-"""
 from slackclient import SlackClient
 
 
 class Handler(object):
+    """
+    here is where the slack handler will go.
+    """
+
     def __init__(self, config):
+        """
+
+        """
         self.slack_token = config["slack"]["token"]
         self.bot_id = str(config["slack"]["bot_id"])