+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.1.2] - 2020-04-04
+### Changes
+    - Added a logic that automatically handles manifest file directory path
+    - Move default services.yaml in config directory
+    - Move manifests folder to hamb package directory
+
 ## [0.1.1] - 2020-04-03
 ### Changes
     - Fix hamb script execution
@@ -15,5 +21,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
     - Initial class import
 
+[0.1.2]: https://github.com/equinoxfitness/hamb/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/equinoxfitness/hamb/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/equinoxfitness/hamb/releases/tag/v0.1.0